feat(node-service): Refactor `EngineActor` (#1867)
## Overview

Refactors the `EngineActor` a bit, thinning out the control loop and
spreading out the module.
