+++ crates/node/service/src/actors/engine/error.rs
@@ -1,0 +1,21 @@
+//! Error type for the [`EngineActor`].
+//!
+//! [`EngineActor`]: super::EngineActor
+
+use kona_engine::{EngineResetError, EngineTaskError};
+
+/// An error from the [`EngineActor`].
+///
+/// [`EngineActor`]: super::EngineActor
+#[derive(thiserror::Error, Debug)]
+pub enum EngineError {
+    /// Closed channel error.
+    #[error("a channel has been closed unexpectedly")]
+    ChannelClosed,
+    /// Engine reset error.
+    #[error(transparent)]
+    EngineReset(#[from] EngineResetError),
+    /// Engine task error.
+    #[error(transparent)]
+    EngineTask(#[from] EngineTaskError),
+}
