+++ crates/node/service/src/actors/engine/mod.rs
@@ -1,0 +1,10 @@
+//! The [`EngineActor`] and its components.
+
+mod actor;
+pub use actor::{EngineActor, EngineLauncher, InboundEngineMessage};
+
+mod error;
+pub use error::EngineError;
+
+mod finalizer;
+pub use finalizer::L2Finalizer;
