+++ crates/node/service/src/lib.rs
@@ -18,8 +18,9 @@ pub use service::{
 mod actors;
 pub use actors::{
     DerivationActor, DerivationError, EngineActor, EngineError, EngineLauncher,
-    InboundDerivationMessage, L1WatcherRpc, L1WatcherRpcError, NetworkActor, NetworkActorError,
-    NodeActor, RpcActor, RpcActorError, RuntimeActor, RuntimeLauncher,
+    InboundDerivationMessage, InboundEngineMessage, L1WatcherRpc, L1WatcherRpcError, L2Finalizer,
+    NetworkActor, NetworkActorError, NodeActor, RpcActor, RpcActorError, RuntimeActor,
+    RuntimeLauncher,
 };
 
 mod metrics;
